% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_pbp.R
\name{basketball_pbp}
\alias{basketball_pbp}
\title{Returns usports basketball play by play data}
\usage{
basketball_pbp(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports basketball play by play data
}
\examples{
basketball_pbp("m")
basketball_pbp("w")
}
