% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_team_box_score.R
\name{basketball_team_box_score}
\alias{basketball_team_box_score}
\title{Returns usports basketball team box scores}
\usage{
basketball_team_box_score(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports basketball team box scores
}
\examples{
basketball_team_box_score("m")
basketball_team_box_score("w")
}
