% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_hockey_pbp.R
\name{field_hockey_pbp}
\alias{field_hockey_pbp}
\title{Get USPORTS field hockey pbp}
\usage{
field_hockey_pbp(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS field hockey pbp
}
\examples{
\dontrun{field_hockey_pbp(2018)}
}
