% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_hockey_player_box_score.R
\name{field_hockey_player_box_score}
\alias{field_hockey_player_box_score}
\title{Returns usports field hockey player box scores}
\usage{
field_hockey_player_box_score()
}
\value{
A Dataframe.
}
\description{
Returns usports field hockey player box scores
}
\examples{
field_hockey_player_box_score()
}
