% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_hockey_schedule.R
\name{field_hockey_schedule}
\alias{field_hockey_schedule}
\title{Returns usports field hockey schedule}
\usage{
field_hockey_schedule()
}
\value{
A Dataframe.
}
\description{
Returns usports field hockey schedule
}
\examples{
field_hockey_schedule()
}
