% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_hockey_team_box_score.R
\name{field_hockey_team_box_score}
\alias{field_hockey_team_box_score}
\title{Returns usports field hockey team box scores}
\usage{
field_hockey_team_box_score(year = c())
}
\value{
A Dataframe.
}
\description{
Returns usports field hockey team box scores
}
\examples{
field_hockey_team_box_score()
}
