% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_drive_summaries.R
\name{football_drive_summaries}
\alias{football_drive_summaries}
\title{Returns usports football drive summaries}
\usage{
football_drive_summaries(year = c())
}
\value{
A Dataframe.
}
\description{
Returns usports football drive summaries
}
\examples{
football_drive_summaries()
}
