% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_pbp.R
\name{football_pbp}
\alias{football_pbp}
\title{Get USPORTS football pbp}
\usage{
football_pbp(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS football pbp
}
\examples{
\dontrun{football_pbp(2018)}
}
