% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_returns.R
\name{football_returns}
\alias{football_returns}
\title{Get USPORTS football returns stats}
\usage{
football_returns(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS football returns stats
}
\examples{
football_returns(2018)
}
