% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_schedule.R
\name{football_schedule}
\alias{football_schedule}
\title{Get USPORTS football schedule}
\usage{
football_schedule()
}
\value{
A data frame.
}
\description{
Get USPORTS football schedule
}
\examples{
football_schedule()
}
