% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_player_box_score.R
\name{hockey_player_box_score}
\alias{hockey_player_box_score}
\title{Returns usports hockey player box scores}
\usage{
hockey_player_box_score(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A Dataframe.
}
\description{
Returns usports hockey player box scores
}
\examples{
hockey_player_box_score("m",2018)
hockey_player_box_score("w",2018)
}
