% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_team_box_score.R
\name{hockey_team_box_score}
\alias{hockey_team_box_score}
\title{Returns usports hockey team box scores}
\usage{
hockey_team_box_score(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports hockey team box scores
}
\examples{
hockey_team_box_score("m")
hockey_team_box_score("w")
}
