% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv}
\alias{read_csv}
\title{reads a csv from a url with a csv}
\usage{
read_csv(file_url)
}
\arguments{
\item{file_url}{A URL to a CSV file.}
}
\value{
A data frame with the first few rows of the CSV file.
}
\description{
reads a csv from a url with a csv
}
