% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rugby_schedules.R
\name{rugby_schedule}
\alias{rugby_schedule}
\title{Get USPORTS rugby schedule}
\usage{
rugby_schedule(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A tibble.
}
\description{
Get USPORTS rugby schedule
}
\examples{
rugby_schedule("m")
rugby_schedule("w")
}
