% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer_pbp.R
\name{soccer_pbp}
\alias{soccer_pbp}
\title{Get USPORTS soccer play by play data}
\usage{
soccer_pbp(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS soccer play by play data
}
\examples{
soccer_pbp("m",2018)
soccer_pbp("w",2018)
}
