% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming_athlete_rankings.R
\name{swimming_athlete_rankings}
\alias{swimming_athlete_rankings}
\title{Get U SPORTS swimming athletes rankings}
\usage{
swimming_athlete_rankings()
}
\value{
A tibble.
}
\description{
Get U SPORTS swimming athletes rankings
}
\examples{
swimming_athlete_rankings()
}
