% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming_team_rankings.R
\name{swimming_team_rankings}
\alias{swimming_team_rankings}
\title{Returns usports swimming team rankings}
\usage{
swimming_team_rankings()
}
\value{
A Dataframe.
}
\description{
Returns usports swimming team rankings
}
\examples{
swimming_team_rankings()
}
