% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnf_athlete_rankings.R
\name{tnf_athlete_rankings}
\alias{tnf_athlete_rankings}
\title{Get USPORTS track and field athlete rankings}
\usage{
tnf_athlete_rankings()
}
\value{
A data frame.
}
\description{
Get USPORTS track and field athlete rankings
}
\examples{
tnf_athlete_rankings()
}
