% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnf_meet_results.R
\name{tnf_meet_results}
\alias{tnf_meet_results}
\title{Returns usports track and field meet results}
\usage{
tnf_meet_results()
}
\value{
A Dataframe.
}
\description{
Returns usports track and field meet results
}
\examples{
tnf_meet_results()
}
