% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnf_universities.R
\name{tnf_universities}
\alias{tnf_universities}
\title{Returns Usports track and field universities}
\usage{
tnf_universities()
}
\value{
A Dataframe.
}
\description{
Returns Usports track and field universities
}
\examples{
tnf_universities()
}
