% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyball_pbp.R
\name{volleyball_pbp}
\alias{volleyball_pbp}
\title{Returns usports volleyball play by play data}
\usage{
volleyball_pbp(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports volleyball play by play data
}
\examples{
volleyball_pbp("m")
volleyball_pbp("w")
}
