% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyball_player_box_score.R
\name{volleyball_player_box_score}
\alias{volleyball_player_box_score}
\title{Returns usports volleyball player box scores}
\usage{
volleyball_player_box_score(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports volleyball player box scores
}
\examples{
volleyball_player_box_score("m")
volleyball_player_box_score("w")
}
