% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyball_team_box_score.R
\name{volleyball_team_box_score}
\alias{volleyball_team_box_score}
\title{Returns usports volleyball team box scores}
\usage{
volleyball_team_box_score(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports volleyball team box scores
}
\examples{
volleyball_team_box_score("m")
volleyball_team_box_score("w")
}
