% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrestling_athlete_rankings.R
\name{wrestling_athlete_rankings}
\alias{wrestling_athlete_rankings}
\title{Get USPORTS wrestling athletes rankings}
\usage{
wrestling_athlete_rankings(gender, weight)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{weight}{weight class}
}
\value{
A tibble.
}
\description{
Get USPORTS wrestling athletes rankings
}
\examples{
wrestling_athlete_rankings("m",90)
wrestling_athlete_rankings("w",54)
}
