% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_meet_results.R
\name{xc_meet_results}
\alias{xc_meet_results}
\title{Get USPORTS XC meet results}
\usage{
xc_meet_results()
}
\value{
A data frame.
}
\description{
Get USPORTS XC meet results
}
\examples{
xc_meet_results()
}
