% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_team_rankings.R
\name{xc_team_rankings}
\alias{xc_team_rankings}
\title{Get USPORTS XC Team Rankings}
\usage{
xc_team_rankings()
}
\value{
A tibble.
}
\description{
Get USPORTS XC Team Rankings
}
\examples{
xc_team_rankings()
}
