% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_pbp.R
\name{basketball_pbp}
\alias{basketball_pbp}
\title{Get USPORTS basketball pbp data}
\usage{
basketball_pbp(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS basketball pbp data
}
\examples{
\dontrun{basketball_pbp("m",2018)}
}
