% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_player_box_score.R
\name{basketball_player_box_score}
\alias{basketball_player_box_score}
\title{Returns usports basketball players box scores}
\usage{
basketball_player_box_score(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A Dataframe.
}
\description{
Returns usports basketball players box scores
}
\examples{
basketball_player_box_score("m",2018)
basketball_player_box_score("w",2018)
}
