% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_schedule.R
\name{basketball_schedule}
\alias{basketball_schedule}
\title{Get USPORTS basketball schedule}
\usage{
basketball_schedule(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A data frame.
}
\description{
Get USPORTS basketball schedule
}
\examples{
basketball_schedule("m")
}
