% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_hockey_player_box_score.R
\name{field_hockey_player_box_score}
\alias{field_hockey_player_box_score}
\title{Get USPORTS field hockey player box scores}
\usage{
field_hockey_player_box_score(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS field hockey player box scores
}
\examples{
field_hockey_player_box_score(2018)
}
