% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_hockey_schedule.R
\name{field_hockey_schedule}
\alias{field_hockey_schedule}
\title{Get USPORTS field hockey schedule}
\usage{
field_hockey_schedule()
}
\value{
A data frame.
}
\description{
Get USPORTS field hockey schedule
}
\examples{
field_hockey_schedule()
}
