% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_defence.R
\name{football_defence}
\alias{football_defence}
\title{Get USPORTS football defence stats}
\usage{
football_defence(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A tibble.
}
\description{
Get USPORTS football defence stats
}
\examples{
football_defence(2018)
}
