% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_drive_summaries.R
\name{football_drive_summaries}
\alias{football_drive_summaries}
\title{Get USPORTS football drive summaries}
\usage{
football_drive_summaries(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS football drive summaries
}
\examples{
football_drive_summaries(2018)
}
