% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_kicking.R
\name{football_kicking}
\alias{football_kicking}
\title{Returns usports football kicking stats}
\usage{
football_kicking(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A Dataframe.
}
\description{
Returns usports football kicking stats
}
\examples{
football_kicking(2018)
}
