% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_offence.R
\name{football_offence}
\alias{football_offence}
\title{Returns usports football offence stats}
\usage{
football_offence(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A Dataframe.
}
\description{
Returns usports football offence stats
}
\examples{
football_offence(2018)
}
