% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_pbp.R
\name{football_pbp}
\alias{football_pbp}
\title{Returns usports football schedule}
\usage{
football_pbp()
}
\value{
A Dataframe.
}
\description{
Returns usports football schedule
}
\examples{
football_pbp()
}
