% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_returns.R
\name{football_returns}
\alias{football_returns}
\title{Returns usports football returns stats}
\usage{
football_returns(year = c())
}
\value{
A Dataframe.
}
\description{
Returns usports football returns stats
}
\examples{
football_returns()
}
