% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_schedule.R
\name{football_schedule}
\alias{football_schedule}
\title{Returns usports football schedule}
\usage{
football_schedule()
}
\value{
A Dataframe.
}
\description{
Returns usports football schedule
}
\examples{
football_schedule()
}
