% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_scoring_summaries.R
\name{football_scoring_summaries}
\alias{football_scoring_summaries}
\title{Returns usports football scoring summaries}
\usage{
football_scoring_summaries(year = c())
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A Dataframe.
}
\description{
Returns usports football scoring summaries
}
\examples{
football_scoring_summaries(2018)
}
