% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_pbp.R
\name{hockey_pbp}
\alias{hockey_pbp}
\title{Returns usports hockey play by play data}
\usage{
hockey_pbp(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports hockey play by play data
}
\examples{
hockey_pbp("m")
hockey_pbp("w")
}
