% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_pbp.R
\name{hockey_pbp}
\alias{hockey_pbp}
\title{Get USPORTS hockey pbp}
\usage{
hockey_pbp(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A tibble.
}
\description{
Get USPORTS hockey pbp
}
\examples{
hockey_pbp("m",2018)
}
