% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_schedule.R
\name{hockey_schedule}
\alias{hockey_schedule}
\title{Get USPORTS hockey schedules}
\usage{
hockey_schedule(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A data frame.
}
\description{
Get USPORTS hockey schedules
}
\examples{
hockey_schedule("m")
}
