% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_schedule.R
\name{hockey_schedule}
\alias{hockey_schedule}
\title{Returns usports hockey schedules}
\usage{
hockey_schedule(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports hockey schedules
}
\examples{
hockey_schedule("m")
hockey_schedule("w")
}
