% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_team_box_score.R
\name{hockey_team_box_score}
\alias{hockey_team_box_score}
\title{Get USPORTS hockey team box scores}
\usage{
hockey_team_box_score(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS hockey team box scores
}
\examples{
hockey_team_box_score("m",2018)
hockey_team_box_score("w",2018)
}
