% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer_pbp.R
\name{soccer_pbp}
\alias{soccer_pbp}
\title{Returns usports soccer play by play data}
\usage{
soccer_pbp(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports soccer play by play data
}
\examples{
soccer_pbp("m")
soccer_pbp("w")
}
