% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer_schedule.R
\name{soccer_schedule}
\alias{soccer_schedule}
\title{Get USPORTS soccer schedule}
\usage{
soccer_schedule(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A data frame.
}
\description{
Get USPORTS soccer schedule
}
\examples{
soccer_schedule("m")
soccer_schedule("w")
}
