% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming_athlete_rankings.R
\name{swimming_athlete_rankings}
\alias{swimming_athlete_rankings}
\title{Returns usports swimming athletes rankings}
\usage{
swimming_athlete_rankings()
}
\value{
A Dataframe.
}
\description{
Returns usports swimming athletes rankings
}
\examples{
swimming_athlete_rankings()
}
