% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnf_rosters.R
\name{tnf_rosters}
\alias{tnf_rosters}
\title{Get U SPORTS track and field rosters}
\usage{
tnf_rosters()
}
\value{
A tibble.
}
\description{
Get U SPORTS track and field rosters
}
\examples{
tnf_rosters()
}
