% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnf_team_rankings.R
\name{tnf_team_rankings}
\alias{tnf_team_rankings}
\title{Get U SPORTS track and field team rankings}
\usage{
tnf_team_rankings()
}
\value{
A tibble.
}
\description{
Get U SPORTS track and field team rankings
}
\examples{
tnf_team_rankings()
}
