% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyball_pbp.R
\name{volleyball_pbp}
\alias{volleyball_pbp}
\title{Get USPORTS volleyball play by play data}
\usage{
volleyball_pbp(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A tibble.
}
\description{
Get USPORTS volleyball play by play data
}
\examples{
volleyball_pbp("m",2018)
}
