% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyball_player_box_score.R
\name{volleyball_player_box_score}
\alias{volleyball_player_box_score}
\title{Get USPORTS volleyball player box scores}
\usage{
volleyball_player_box_score(gender, year = c())
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}

\item{year}{The season(s) of interest.}
}
\value{
A data frame.
}
\description{
Get USPORTS volleyball player box scores
}
\examples{
volleyball_player_box_score("m",2018)
}
