% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyball_schedule.R
\name{volleyball_schedule}
\alias{volleyball_schedule}
\title{Get USPORTS volleyball schedule}
\usage{
volleyball_schedule(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A tibble.
}
\description{
Get USPORTS volleyball schedule
}
\examples{
volleyball_schedule("m")
}
