% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrestling_team_rankings.R
\name{wrestling_team_rankings}
\alias{wrestling_team_rankings}
\title{Returns usports wrestling team rankings}
\usage{
wrestling_team_rankings(gender)
}
\arguments{
\item{gender}{Gender of athletes "m" or "w"}
}
\value{
A Dataframe.
}
\description{
Returns usports wrestling team rankings
}
\examples{
wrestling_team_rankings("m")
wrestling_team_rankings("w")
}
