% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_meet_results.R
\name{xc_meet_results}
\alias{xc_meet_results}
\title{Returns Usports XC meet results}
\usage{
xc_meet_results()
}
\value{
A Dataframe.
}
\description{
Returns Usports XC meet results
}
\examples{
xc_meet_results()
}
